create rule tranlator service to avoid repeated code
